+++ src/components/FormConstructor/FormConstructorFormBlockConfigure/Elements/ButtonFormElement/index.ts
@@ -1,0 +1,1 @@
+export { ButtonFormElement } from './ButtonFormElement'
