+++ src/components/FormConstructor/FormConstructorFormBlockConfigure/Elements/LayoutFromElement/index.ts
@@ -1,0 +1,1 @@
+export { LayoutFromElement } from './LayoutFormElement'
