+++ src/components/FormConstructor/FormConstructorFormBlockConfigure/FormConstructorFormBlockConfigure.tsx
@@ -1,16 +1,14 @@
 import React, { FC, ReactNode } from 'react'
 import { Provider } from 'react-redux'
-import { ConpomonentStructure } from './Panels/ConpomonentStructure'
-import { Settings } from './Panels/SettingsBlock'
 import { store } from '../store'
 import { IFormConstructorFormBlockConfigure } from './types'
-import { FormBlock } from './Panels/FormBlock'
+import { ConpomonentStructure, FormBlock, Settings } from './Panels'
 
 interface Props {
   children?: ReactNode
 }
 
-const FormConstructorFormBlockConfigure: FC<Props> & IFormConstructorFormBlockConfigure = ({
+export const FormConstructorFormBlockConfigure: FC<Props> & IFormConstructorFormBlockConfigure = ({
   children,
 }) => {
   return <Provider store={store}>{children}</Provider>
@@ -19,5 +17,3 @@ const FormConstructorFormBlockConfigure: FC<Props> & IFormConstructorFormBlockCo
 FormConstructorFormBlockConfigure.Settings = Settings
 FormConstructorFormBlockConfigure.WhiteFormBlock = FormBlock
 FormConstructorFormBlockConfigure.ConpomonentStructure = ConpomonentStructure
-
-export default FormConstructorFormBlockConfigure
