+++ src/components/FormConstructor/FormConstructorFormBlockConfigure/Panels/ConpomonentStructure/ComponentItems/ComponentCard/index.ts
@@ -1,0 +1,4 @@
+import { ComponentCard } from './ComponentCard'
+import { IComponentCard } from './types'
+
+export { ComponentCard, type IComponentCard }
