+++ src/components/FormConstructor/FormConstructorFormBlockConfigure/Panels/ConpomonentStructure/ComponentItems/index.ts
@@ -1,0 +1,1 @@
+export { ComponentItems } from './ComponentItems'
