+++ src/components/FormConstructor/FormConstructorFormBlockConfigure/Panels/ConpomonentStructure/ComponentStructure.tsx
