+++ src/components/FormConstructor/FormConstructorFormBlockConfigure/Panels/ConpomonentStructure/ComponentTree/index.ts
@@ -1,0 +1,1 @@
+export { ComponentTree } from './ComponentTree'
