+++ src/components/FormConstructor/FormConstructorFormBlockConfigure/Panels/ConpomonentStructure/index.ts
@@ -1,0 +1,1 @@
+export { ConpomonentStructure } from './ComponentStructure'
