+++ src/components/FormConstructor/FormConstructorFormBlockConfigure/Panels/FormBlock/index.ts
@@ -1,0 +1,1 @@
+export { FormBlock } from './FormBlock'
