+++ src/components/FormConstructor/FormConstructorFormBlockConfigure/Panels/Settings/LayoutSettings/index.ts
@@ -1,0 +1,1 @@
+export { LayoutSettings } from './LayoutSettings'
