+++ src/components/FormConstructor/FormConstructorFormBlockConfigure/Panels/Settings/index.ts
@@ -1,0 +1,1 @@
+export { Settings } from './Settings'
