+++ src/components/FormConstructor/FormConstructorFormBlockConfigure/Panels/index.ts
@@ -1,0 +1,5 @@
+import { FormBlock } from './FormBlock'
+import { ConpomonentStructure } from './ConpomonentStructure'
+import { Settings } from './Settings'
+
+export { Settings, FormBlock, ConpomonentStructure }
