+++ src/components/FormConstructor/FormConstructorFormBlockConfigure/SelectableLayer/index.ts
@@ -1,0 +1,1 @@
+export { SelectableLayer } from './SelectableLayer'
