+++ src/components/FormConstructor/FormConstructorFormBlockConfigure/index.ts
@@ -1,0 +1,1 @@
+export { FormConstructorFormBlockConfigure } from './FormConstructorFormBlockConfigure'
